requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
PyYAML>=6.0.1
feedparser>=6.0.11
APScheduler>=3.10,<4
SQLAlchemy>=2.0.29
alembic>=1.13.2
psycopg[binary]>=3.1.18
